requirements.txt
### libraries for co2 sensor project

# general
adafruit_bus_device

# scd_30 co2 sensor
adafruit_scd30.mpy

# airlift featherwing
adafruit_esp32spi
adafruit_io
adafruit_minimqtt
adafruit_register
adafruit_requests.mpy

# display
adafruit_bitmap_font
adafruit_displayio_ssd1306.mpy
adafruit_display_shapes
adafruit_display_text
neopixel.mpy
